Refactor Fisher
Refactors Fisher into a class with related functions.

To access the Fisher information matrix, a Fisher object can be created the usual way with `Fisher([qs], model.xi, counts, [model])`.

The Fisher information matrix is an attribute of this Class, defined by `g`. The reasoning for having it defined as a property, is that this makes this attribute self-updating. E.g. if we were to change anything in one of the parameters (for example we want to know what it would be with different counting times), we don't need to re-initialize the class, but can just change the attribute directly and call g again. e.g. 'g = fisher.g`. I'm open to a more verbose name. (e.g. `fisher.information_matrix`).

The eigenvalue is defined in a similar way, and can be accessed from the same Fisher object: `fisher.min_eigenval`.

I tried to split the main method such that they all have a single responsibility (the first S in the SOLID design principles). Due to the self-updating nature of `g` and `min_eigenval` properties, the other methods should not really be accessed in user-space.